ci-bench-runner: avoid OpenSSL dependency
The `sentry` crate uses `native-tls` by default, which requires
`openssl` and `pkg-config` be installed. Let's use Rustls instead by
disabling default features and specifying the `rustls` feature
explicitly.

We also have to choose an explicit transport to use with `rustls`, this
commit chooses `ureq` as a simple, pure-rust transport that supports
`rustls`.

The net result is we no longer need to have `pkg-config` or `openssl`
available to build the application.