ansible: build ci-bench-runner on server
Previously the `ci-bench-runner` application had to be build in
a release configuration locally, on the deployment machine. The Ansible
deploy role would then copy the binary (which must be created ahead of
time) to the remote system.

This has the advantage of not needing to do an expensive build operation
on the benchmarking server, where it may introduce noise, but has a few
important disadvantages:

* The deployment machine and the benchmarking machine must be the same
  architecture
* The deployment machine and the benchmarking machine must have the same
  dynamic library paths

This commit replaces this approach with a task that builds
`ci-bench-runner` on the benchmarking server. This resolves these
disadvantages.

We'll need to take care not to run the deployment scripts when
benchmarking is on the go, so a note is added to the Ansible README.

The new tasks are tagged to match the existing deploy role: we use
`setup` for each step within the build task, and `deploy` for the
top-level `playbook.yml` part. In a subsequent commit we will rename
these tags to make it clearer it's both a build-and-deploy we
want to perform together