Discard events that caused a crash, instead of retrying
Retrying makes sense for transient errors, but not when errors are
permanent. In the latter case, the previous code would enter an infinite
crash loop, constantly retrying to handle an event that would result in
an error.

Faced with the option to come up with some system to limit the amount of
retries, we re-evaluated our strategy and decided to avoid retrying at
all.