github/job: use id to detect our own comment, not login
Previously the `handle_issue_comment` logic used the GitHub login string
of the user posting the comment to decide if it was a comment from the
bot. This requires hardcoding the name of the application, which may be
fragile.

Instead, include the ID from the GitHub user model and in
`handle_issue_comment` check if it matches the configured application
GitHub application ID.