flake: remove package definition
A package derivation isn't very useful for this project. The deployment
model assumes a Cargo based build on the remote server, orchestrated by
Ansible. For local development we also prefer Cargo. Having the package
derivation would be useful if we wanted to deploy with Nix, but we
don't. It also makes taking Git dependencies in our Cargo.toml
cumbersome, and may break builds of tools like bencher-client that
aren't written with an isolated build environment in mind.